requirements.txt
discord
python-dotenv
datetime
pandas